% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ukb_brainmri_plots.R
\name{create_test_plot}
\alias{create_test_plot}
\title{Title: Plotting functions for rejection rates of CITs}
\usage{
create_test_plot(
  folder_path,
  dncits,
  settings,
  sample_sizes = NULL,
  betas = NULL
)
}
\arguments{
\item{folder_path}{path to folder with rejection rates of all settings and all CITs}

\item{dncits}{vector of all DNCITs}

\item{settings}{data frame with all settings, columns should consist of settings$confounder, fct_relation, embedding_map, noise_x}

\item{sample_sizes}{vector of all sample sizes (either sample_sizes or betas has to be specified)}

\item{betas}{vector of all betas (either betas or sample_sizes has to be specified)}
}
\value{
returns a plot with all rejection rates of all settings and all CITs
}
\description{
Title: Plotting functions for rejection rates of CITs
}
