% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_ixi.R
\name{data_gen_ixi}
\alias{data_gen_ixi}
\title{IXI Data Generation with Neural Network Embeddings}
\usage{
data_gen_ixi(
  seed,
  idx_sample = NULL,
  n_sample = NULL,
  idx_beta2 = NULL,
  beta2s = NULL,
  post_non_lin = 4,
  eps_sigmaX = 0,
  eps_sigmaY = 1,
  eps_sigmaZ = 0,
  embedding_orig = "medicalnet",
  embedding_obs = "scratch",
  confounder = "ukb_z4",
  g_z = "linear"
)
}
\arguments{
\item{seed}{random seed for current data generation}

\item{idx_sample}{index of current sample size}

\item{n_sample}{list of sample sizes}

\item{idx_beta2}{index of current beta2}

\item{beta2s}{list of beta2s}

\item{post_non_lin}{post non-linearity function}

\item{eps_sigmaX}{noise level for X}

\item{eps_sigmaY}{noise level for Y}

\item{eps_sigmaZ}{noise level for Z}

\item{embedding_orig}{embedding map used to obtain original feature representations (used to simulate conditional association or CI)}

\item{embedding_obs}{embedding map used to obtain observed feature representations (used for CI testing)}

\item{confounder}{confounder used to simulate Y}

\item{g_z}{confounder response relationship}

\item{n}{sample size}
}
\value{
list of X_obs, Y, Z
}
\description{
IXI Data Generation with Neural Network Embeddings
}
