% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dncitPaper-package.R
\docType{package}
\name{dncitPaper-package}
\alias{dncitPaper}
\alias{dncitPaper-package}
\title{dncitPaper: Code to reproduce the analysis of the paper Deep Nonparametric Conditional Indpendence Tests (DNCIT)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Marco Simnacher \email{marco.simnacher@hu-berlin.de}

}
\keyword{internal}
