% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_diagnostic.R
\name{is_binary}
\alias{is_binary}
\title{Check if a column is binary}
\usage{
is_binary(x)
}
\arguments{
\item{x}{Vector to check}
}
\value{
Logical indicating if x has only 2 unique values
}
\description{
Check if a column is binary
}
