% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_diagnostic.R
\name{load_Z}
\alias{load_Z}
\title{Load confounders Z based on confounder type}
\usage{
load_Z(path_to_ukb_data, confounder)
}
\arguments{
\item{path_to_ukb_data}{Path to UKB data directory}

\item{confounder}{Type of confounder to load}
}
\value{
Data frame with confounders including id column
}
\description{
Load confounders Z based on confounder type
}
