% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ukb_brainmri_plots.R
\name{sim_ukb_brainmri_plots}
\alias{sim_ukb_brainmri_plots}
\title{Title: Plotting functions for rejection rates of DNCITs}
\usage{
sim_ukb_brainmri_plots()
}
\value{
returns two plots with all rejection rates of all settings and all DNCITs
}
\description{
change manually and run line by line
}
