% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_from_xz.R
\name{y_from_xz}
\alias{y_from_xz}
\title{Simulate response Y from Z and optionally X}
\usage{
y_from_xz(
  Z,
  eps_sigmaY,
  X = NULL,
  beta2s = NULL,
  idx_beta2 = NULL,
  post_non_lin = 1
)
}
\arguments{
\item{Z}{observed confounders}

\item{eps_sigmaY}{integer of (homogeneous) standard deviation of noise in Y}

\item{X}{(non-observed) feature representation used to simulate response}

\item{post_non_lin}{integer specifying nonlinear function g() applied to input consisting of X, Z and epsY}
}
\value{
returns function to construct Y
}
\description{
Simulate response Y from Z and optionally X
}
